+++ lib/jsapi/controller/authentication/credentials/api_key.rb
@@ -4,17 +4,18 @@ module Jsapi
   module Controller
     module Authentication
       module Credentials
-        # API key credentials
+        # Holds an API key.
         class APIKey
+          # The API key as it was passed.
           attr_reader :api_key
 
           def initialize(api_key)
             @api_key = api_key
           end
 
-          # Returns true if the API key is well formed.
+          # Returns true if the API Key is not nil.
           def well_formed?
-            api_key.present?
+            !api_key.nil?
           end
         end
       end