+++ lib/jsapi/controller/methods/callbacks/class_methods.rb
@@ -11,16 +11,16 @@ module Jsapi
           #   api_before_processing(method, except: nil, only: nil)
           #   api_before_processing(except: nil, only: nil, &block)
           #
-          # Triggered by +api_operation+ and +api_operation!+ before the operation is
-          # processed.
+          # Registers a callback that is triggered by +api_operation+ and
+          # +api_operation!+ before the operation is processed.
           #
-          #   api_before_processing do |api_params
+          #   api_before_processing do |api_params|
           #     # ...
           #   end
           #
-          # An +api_before_processing+ callback must be able to accept one positional
-          # argument to receive the request parameters. The returned value is not
-          # used in further processing.
+          # An +api_before_processing+ callback must be able to accept one
+          # positional argument to receive the request parameters. The
+          # returned value is not used in further processing.
 
           ##
           # :method: api_before_rendering
@@ -28,17 +28,18 @@ module Jsapi
           #   api_before_rendering(method, except: nil, only: nil)
           #   api_before_rendering(except: nil, only: nil, &block)
           #
-          # Triggered by +api_operation+ and +api_operation!+ before the response body
-          # is rendered.
+          # Registers a callback that is triggered by +api_operation+ and
+          # +api_operation!+ before the response body is rendered.
           #
           #   api_before_rendering do |result, api_params|
           #     { request_id: api_params.request_id, payload: result }
           #   end
           #
-          # An +api_before_rendering+ callback must be able to accept two positional
-          # arguments to receive the result to be rendered as the first argument and
-          # the request parameters as the second argument. Instead of the passed
-          # result, the returned value is rendered.
+          # An +api_before_rendering+ callback must be able to accept two
+          # positional arguments to receive the result to be rendered as the
+          # first argument and the request parameters as the second argument.
+          # The value returned by the callback replaces the result to be
+          # rendered.
 
           %i[before_processing before_rendering].each do |kind|
             define_method(:"api_#{kind}") do |method = nil, **options, &block|
@@ -46,7 +47,8 @@ module Jsapi
               raise ArgumentError, 'either a method or a block must be ' \
                                    'specified' unless method_or_block
 
-              ((@_api_callbacks ||= {})[kind] ||= []) << Callback.new(method_or_block, **options)
+              ((@_api_callbacks ||= {})[kind] ||= []) <<
+                Callback.new(method_or_block, **options)
             end
           end